functions/requirements.txt
flask==2.0.1
gunicorn==20.1.0
functions-framework
PyMuPDF
mammoth
Werkzeug
python-docx
